aperture routines moeved to `aperture_utils.py`
