+++ src/psfmachine/utils.py
@@ -388,60 +388,3 @@ def sparse_lessthan(arr, limit):
         shape=arr.shape,
     ).astype(bool)
     return masked_arr
-
-
-def compute_FLFRCSAP(psf_models, aperture_mask):
-    """
-    Compute fraction of target flux enclosed in the optimal aperture to total flux
-    for a given source (flux completeness).
-    Follows definition by Kinemuchi at al. 2012.
-    Parameters
-    ----------
-    psf_models : scipy.sparce.csr_matrix
-        Sparse matrix with the PSF models for all targets in the scene. It has shape
-        [n_sources, n_pixels].
-    aperture_mask: numpy.ndarray
-        Array of boolean indicating the aperture for the target source. It has shape of
-        [n_sources, n_pixels].
-
-    Returns
-    -------
-    FLFRCSAP: numpy.ndarray
-        Completeness metric
-    """
-    return np.array(
-        psf_models.multiply(aperture_mask.astype(float)).sum(axis=1)
-        / psf_models.sum(axis=1)
-    ).ravel()
-
-
-def compute_CROWDSAP(psf_models, aperture_mask, idx=None):
-    """
-    Compute the ratio of target flux relative to flux from all sources within
-    the photometric aperture (i.e. 1 - Crowdeness).
-    Follows definition by Kinemuchi at al. 2012.
-    Parameters
-    ----------
-    psf_models : scipy.sparce.csr_matrix
-        Sparse matrix with the PSF models for all targets in the scene. It has shape
-        [n_sources, n_pixels].
-    aperture_mask : numpy.ndarray
-        Array of boolean indicating the aperture for the target source. It has shape of
-        [n_sources, n_pixels].
-    idx : int
-        Source index for what the metric is computed. Value has to be betweeen 0 and
-        psf_model first dimension size.
-        If None, it returns the metric for all sources (first dimension of psf_model).
-
-    Returns
-    -------
-    CROWDSAP : numpy.ndarray
-        Crowdeness metric
-    """
-    ratio = psf_models.multiply(1 / psf_models.sum(axis=0)).tocsr()
-    if idx is None:
-        return np.array(
-            ratio.multiply(aperture_mask.astype(float)).sum(axis=1)
-        ).ravel() / aperture_mask.sum(axis=1)
-    else:
-        return ratio[idx].toarray()[0][aperture_mask].sum() / aperture_mask.sum()
